+++ tests/testthat/test_plots.R
@@ -8,7 +8,7 @@ context("plots")
 skip_if(!requireNamespace("ggplot2", quietly = TRUE))
 skip_if(!requireNamespace("patchwork", quietly = TRUE))
 
-plot_majority <- segplot(schools00, "race", "school", weight = "n", order = "majority")
+plot_majority <- segplot(schools00, "race", "school", weight = "n", order = "majority", hline = "white")
 plot_majority_fixed <- segplot(schools00, "race", "school", weight = "n", order = "majority_fixed")
 plot_seg <- segplot(schools00, "race", "school", weight = "n", order = "segregation")
 plot_entropy <- segplot(schools00, "race", "school",