+++ NEWS.md
@@ -3,8 +3,9 @@
 - various improvements to compression algorithm
 - add dendrogram visualization
 - allow multiple curves in `segcurve` function
-- segplot: add 'secondary_plot' argument
+- segplot: add optional 'secondary_plot' argument
 - segplot: remove 'title' argument
+- segplot: add optional 'hline' argument
 
 # segregation 1.0.0